requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
-e git+https://github.com/UiL-OTS-labs/django-shared-core.git@v2.0.9#egg=uil-django-core
    # via -r requirements.in
asgiref==3.5.0
    # via django
blessings==1.7
    # via curtsies
bpython==0.22.1
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via pip-tools
closure==20191111
    # via uil-django-core
cryptography==36.0.2
    # via -r requirements.in
curtsies==0.3.10
    # via bpython
cwcwidth==0.1.6
    # via
    #   bpython
    #   curtsies
django==3.2.13
    # via
    #   -r requirements.in
    #   django-auth-ldap
    #   django-axes
    #   django-braces
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-simple-menu
    #   djangorestframework
    #   uil-django-core
django-auth-ldap==4.0.0
    # via -r requirements.in
django-axes==5.32.0
    # via -r requirements.in
django-braces==1.15.0
    # via -r requirements.in
django-csp==3.7
    # via -r requirements.in
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-ipware==4.0.2
    # via django-axes
django-simple-menu==1.2.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   uil-django-core
greenlet==1.1.2
    # via bpython
idna==3.3
    # via requests
lesscpy==0.15.0
    # via uil-django-core
mysqlclient==2.1.0
    # via -r requirements.in
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
ply==3.11
    # via lesscpy
pscript==0.7.7
    # via vbuild
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via bpython
pyjwt==2.3.0
    # via
    #   -r requirements.in
    #   uil-django-core
pyscss==1.4.0
    # via uil-django-core
python-dateutil==2.8.2
    # via -r requirements.in
python-ldap==3.4.0
    # via django-auth-ldap
pytz==2022.1
    # via
    #   django
    #   djangorestframework
pyxdg==0.27
    # via bpython
requests==2.27.1
    # via
    #   bpython
    #   uil-django-core
six==1.16.0
    # via
    #   blessings
    #   lesscpy
    #   pyscss
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via pep517
typing-extensions==4.2.0
    # via bpython
urllib3==1.26.9
    # via requests
vbuild==0.8.1
    # via uil-django-core
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
